// src/api/subformNames.ts
export enum RequiredSubformsEnum {
  AccessToUnitForm = "AccessToUnitForm",
  AccountOwnerForm = "AccountOwnerForm",
  ActivationForm = "ActivationForm",
  AdvancedBookingSettingsForm = "AdvancedBookingSettingsForm",
  BankingForm = "BankingForm",
  BathroomsForm = "BathroomsForm",
  BedroomsForm = "BedroomsForm",
  BuildingAmenitiesForm = "BuildingAmenitiesForm",
  BusinessLicenceForm = "BusinessLicenceForm",
  CalendarBlockOffsForm = "CalendarBlockOffsForm",
  CancellationPoliciesForm = "CancellationPoliciesForm",
  CheckInOutForm = "CheckInOutForm",
  CleaningProtocolForm = "CleaningProtocolForm",
  DamageDepositForm = "DamageDepositForm",
  EmergencyContactsForm = "EmergencyContactsForm",
  EssentialAmenitiesForm = "EssentialAmenitiesForm",
  KitchenAmenitiesForm = "KitchenAmenitiesForm",
  LocationForm = "LocationForm",
  MembershipForm = "MembershipForm",
  NewListingOfferForm = "NewListingOfferForm",
  NonResidentStatusForm = "NonResidentStatusForm",
  ParkingForm = "ParkingForm",
  PartnersForm = "PartnersForm",
  PhotosForm = "PhotosForm",
  PrivateSafetyAmenitiesForm = "PrivateSafetyAmenitiesForm",
  PropertyAddressForm = "PropertyAddressForm",
  PropertyContactForm = "PropertyContactForm",
  RateGroupForm = "RateGroupForm",
  RateSettingsForm = "RateSettingsForm",
  RatesScheduleForm = "RatesScheduleForm",
  ReservationSettingsInstantBookingForm = "ReservationSettingsInstantBookingForm",
  ReservationSettingsPaymentScheduleForm = "ReservationSettingsPaymentScheduleForm",
  SetupQuestionsForm = "SetupQuestionsForm",
  StyleSquareFootageForm = "StyleSquareFootageForm",
  SuitabilitiesForm = "SuitabilitiesForm",
  SurchargesForm = "SurchargesForm",
  TaxesForm = "TaxesForm",
}

export enum OptionalSubformsEnum {
  AllAmenitiesFormOptional = "AllAmenitiesForm",
  PropertyTitleSummaryFormOptional = "PropertyTitleSummaryFormOptional",
  BedroomsFormOptional = "BedroomsFormOptional",
  BathroomsFormOptional = "BathroomsFormOptional",
  LocationFormOptional = "LocationFormOptional",
  EssentialAmenitiesFormOptional = "EssentialAmenitiesFormOptional",
  PrivateSafetyAmenitiesFormOptional = "PrivateSafetyAmenitiesFormOptional",
  KitchenAmenitiesFormOptional = "KitchenAmenitiesFormOptional",
  BuildingAmenitiesFormOptional = "BuildingAmenitiesFormOptional",
  CheckInOutFormOptional = "CheckInOutFormOptional",
  CheckInEmailFormOptional = "CheckInEmailFormOptional",
  ExtendedDescriptionFormOptional = "ExtendedDescriptionFormOptional",
  PropertyPoliciesForm = "PropertyPoliciesForm",
}

export type SubformName = OptionalSubformsEnum | RequiredSubformsEnum;